refactor : 파일구조 변경 add: PopularVoteBodyView생성
